+++ tests/Unit/Public/Get-ScriptLog.Tests.ps1
@@ -1,11 +1,11 @@
 Describe 'Get-ScriptLog' {
-    It 'Requesting the default ScriptLog should fail if no default ScriptLog has been set' {
+    It 'should warn when requesting the default ScriptLog, if no default ScriptLog has been set' {
         Mock -CommandName Write-Warning -MockWith {}
         Get-ScriptLog -Default | Should -BeNullOrEmpty
     }
-    It 'Given no parameters, it should return all ScriptLog objects' {
-        $ScriptLog1 = New-ScriptLog
-        $ScriptLog2 = New-ScriptLog
+    It 'should return all ScriptLog objects, if no parameters are given.' {
+        $ScriptLog1 = New-ScriptLog -LogType Memory
+        $ScriptLog2 = New-ScriptLog -LogType Memory
         $AllScriptLogs = Get-ScriptLog
         $AllScriptLogs.Count | Should -Be 2
     }