+++ CHANGELOG.md
@@ -7,13 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- New function Close-ScriptLog added, to allow for closing active logs during a session.
+
 ### Fixed
 
-- It is no longer possible to have two ScrtipLog objects using the same file.
+- It is no longer possible to have two ScrtipLogs using the same file.
 
 ### Changed
 
-- Added better testing to New-ScriptLog.
+- The parameter AppendDateTime in function New-ScriptLog is now a switch instead of a boolean.
+- Internal test quality improved.
 
 ## [0.3.1] - 2022-07-14